Font and Resource updates + temp Font Patch
- Improved Fonts and Resources (still very much subject to change)
- Fixed a bug pertaining to updating the position of the composable after calculating the font.
- Added a reset (esc) for screens

+++ src/main/kotlin/net/prismclient/aether/ui/composition/Composable.kt
@@ -115,7 +115,7 @@ abstract class Composable(open val modifier: UIModifier<*>) {
     }
 
     open fun composeAnchor() {
-        modifier.anchorPoint?.update(this, width.toFloat(), height.toFloat())
+        modifier.anchorPoint?.update(this, width, height)
     }
 
     open fun composePadding() {
